web: fix useAwait lint errors and decouple retry estimate from DB
Resolve Biome `useAwait` violations by awaiting returned thenables or
dropping `async` where nothing is awaited, and convert flagged object
type aliases to interfaces. Split the pure archive retry-estimate into
`lib/api/archive-retry.ts` so the unit test no longer imports `@/lib/db`
(which throws at import when DATABASE_URL is unset).
